+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Billing: fix error in Document generator order when contact not entered
 - Core: Added new CLI commands to reset the JMAP sync state: https://groupoffice.readthedocs.io/en/latest/system-settings/tools.html
 - Email: case insensitive spoof check
+- Tasks: Split complete and incompleted tasks in links
 
 28-01-2022 6.6.36
 - Tasks: drag and drop tasks to other lists